cp 4d49f95 server: Rework filter_interfaces() for less lambda
The pyflame flamegraphs were highlighting the lambdas as the performance
bottleneck, so break them out procedural style to gain a better
understanding of where the problem lies.

Further, the functional style was almost impenetrable with respect to
the procedural equivalent. When the function we're executing has full
visibility of the transform pipeline the procedural style reduces the
cognitive load (at least, for me) even if it is stateful in some sense.

Change-Id: I444fcc6daedfc7f53932bc7acc950e4f60ab151f
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>

Conflicts:
	obmc/mapper/server.py